Moved the rest of ViewModels to Model folder and removed Vms folder; fixed namespaces.
