+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition/Models/AccountViewModels/SendCodeViewModel.cs
@@ -1,8 +1,8 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
-namespace DevAdventCalendarCompetition.Vms
+namespace DevAdventCalendarCompetition.Models.AccountViewModels
 {
     public class SendCodeViewModel
     {
