+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition/Models/AccountViewModels/VerifyCodeViewModel.cs
@@ -1,7 +1,7 @@
-﻿using System;
+using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace DevAdventCalendarCompetition.Vms
+namespace DevAdventCalendarCompetition.Models.AccountViewModels
 {
     public class VerifyCodeViewModel
     {
