+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition/Models/TestViewModels/TestResultsVm.cs
@@ -1,6 +1,6 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 
-namespace DevAdventCalendarCompetition.Vms
+namespace DevAdventCalendarCompetition.Models.TestViewModels
 {
     public class TestResultsVm
     {
