+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition/Views/Account/SendCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model DevAdventCalendarCompetition.Vms.SendCodeViewModel
+@model DevAdventCalendarCompetition.Models.AccountViewModels.SendCodeViewModel
 @{
     ViewBag.Title = "Wyślij";
 }
