+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition/Views/Account/VerifyCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model DevAdventCalendarCompetition.Vms.VerifyCodeViewModel
+@model DevAdventCalendarCompetition.Models.AccountViewModels.VerifyCodeViewModel
 @{
     ViewBag.Title = "Weryfikacja";
 }
